+++ examples/01_diff_drive_lidar/main.rs
@@ -194,6 +194,7 @@ fn main() {
                 physics: &backend,
                 physics_world,
                 render: None,
+                scene: None,
             },
             &mut data_bus,
         );